front: extract projectedTrains, cutProjectedTrains and cutConflicts in separate files
